move processFile to separate file
